youtube-summarizer-agent/youtube-summarizer-agent/requirements.txt

# Core dependencies
sentient-agent-framework>=0.1.0
python-dotenv>=1.0.0

# YouTube transcript extraction
youtube-transcript-api>=0.6.0

# HTTP client for API calls
httpx>=0.25.0
requests>=2.31.0

# Async support
asyncio>=3.4.3